chore: move recording form types to separate file and hook up schedule type select
